requirements.txt
flask
flask-cors
numpy
opencv-python
pandas
Pillow
rembg
reportlab
gunicorn
openpyxl
